feat: added --no-save and --alarm
Quick update to the code, but in simple terms the -ns or --no-save flag
allow you to prevent your initialised pomodoro from being saved in the
database.

-a or --alarm allow you to pass in a path of an audio file. This will
get played when the pomodoro session is over.

In the future I will work on some more flags to extend the capabilities
of this timer.