feat: break timer after completing a pomodoro
When you completed a pomodoro you will have the option to either quit,
start a new one directly or you can now wait for the break timer!

Now we can actually call this a pomodoro app!
The time for the break is changeable with the -b or --break flag. Pass
in either a value like 1m for minute or 1s for seconds.